+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 7.1.0 (May 14, 2025)
+
+### Added
+- Support for OpenSearch OBO metering for external customers
+- Support for bulk PDB enablement and monitoring in Database Management Service
+### Bug Fix
+- fix infinite loop for OSMH resource discovery
+
 ## 7.0.0 (May 06, 2025)
 
 ### Added